Update category page

+++ resources/js/views/category/index.vue
@@ -3,6 +3,8 @@ import { ref, onMounted } from 'vue';
 import axios from "@/app/http/axios";
 import { useRouter, useRoute } from 'vue-router';
 import {useToast} from 'vue-toast-notification';
+import Swal from 'sweetalert2';
+import draggable from 'vuedraggable';
 
 const toast = useToast();
 const router = useRouter();
@@ -14,9 +16,13 @@ const status = ref(true);
 const issues = ref([]);
 const isShow = ref(false);
 const issue_name = ref('');
+const issue_status = ref(true);
 
 const handleSave = async () => {
     const response = await axios.post(`/api/category/${category_id}`, {name: name.value, status: status.value});
+    issues.value.map(async (item, index) => {
+        await axios.post(`/api/issue/${item.id}`, {order: index+1});
+    });
     if(response) {
         router.push({path: '/edit_workspace', query: {id: workspace_id}});
         toast.success('Category updated successfully.', {position: 'top-right'});
@@ -24,7 +30,7 @@ const handleSave = async () => {
 }
 
 const handleCreate = async () => {
-    const response = await axios.post('/api/issue', {name: issue_name.value, category_id: category_id});
+    const response = await axios.post('/api/issue', {name: issue_name.value, category_id: category_id, status: issue_status.value});
     if(response) {
         toast.success('Create issue successfully', {position: 'top-right'});
         issue_name.value = '';
@@ -34,12 +40,24 @@ const handleCreate = async () => {
 }
 
 const handleDelete = async (id: string) => {
-    const response = await axios.delete(`/api/issue/${id}`);
-    if(response) {
-        issues.value = issues.value.filter(item => item.id != id);
-        toast.success('Issue deleted successfully.', {position: 'top-right'});
-        isShow.value = false;
-    }
+    Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!",
+    }).then(async (result) => {
+        if (result.isConfirmed) {
+            const response = await axios.delete(`/api/issue/${id}`);
+            if(response) {
+                issues.value = issues.value.filter(item => item.id != id);
+            }
+            Swal.fire("Deleted!", "Issue deleted successfully.", "success");
+        }
+    });
+    
 }
 
 onMounted(async () => {
@@ -53,88 +71,95 @@ onMounted(async () => {
 </script>
 
 <template>
-    <v-row justify="end" class="mt-5">
+    <v-row justify="space-between" class="mt-5">
+        <v-btn
+            class="ml-5"
+            color="primary"
+            @click="router.push({path: '/edit_workspace', query: {id: workspace_id}})"
+        >
+            BACK
+        </v-btn>
         <v-btn
             class="mr-5"
-            color="blue"
+            color="primary"
             @click="handleSave()"
         >
             SAVE
         </v-btn>
     </v-row>
     <v-row>
-        <v-col cols="12">
-            <div class="font-weight-medium mt-5">
-                Name <i class="ph-asterisk ph-xs text-danger" />
-            </div>
+        <v-col cols="12" lg="6">
+            <Card title="Edit Category" class="h-100">
+                <v-card-text>
+                    <div class="font-weight-bold mb-2 mt-3">Name <i class="ph-asterisk ph-xs text-danger" /></div>
+                    <TextField
+                        v-model="name"
+                        hide-details
+                        placeholder="Enter category name"
+                    />
+                    <div class="font-weight-bold mb-2 mt-3">Status <i class="ph-asterisk ph-xs text-danger" /></div>
+                    <v-switch
+                        v-model="status"
+                        color="primary"
+                        :label="status ? 'Enable' : 'Disable'"
+                        hide-details
+                        >
+                    </v-switch>
+                    
+                </v-card-text>
+            </Card>
         </v-col>
-        <v-col cols="12">
-            <v-text-field
-                v-model="name"
-                isRequired
-                placeholder="Enter category name"
-                hideDetails
-            ></v-text-field>
-        </v-col>
-        <v-col cols="12">
-            <div class="font-weight-medium mt-5">
-                Status <i class="ph-asterisk ph-xs text-danger" />
-            </div>
-        </v-col>
-        <v-col cols="12">
-            <v-switch
-                v-model="status"
-                color="primary"
-                :label="status ? 'Enable' : 'Disable'"
-                hide-details
-                >
-            </v-switch>
-        </v-col>
-        <v-col col="12">
-            <v-btn
-                @click="isShow = true"
-                color="blue"
-            >
-                Add Issue
-            </v-btn>
-        </v-col>
-        <v-col cols="12">
-            <v-card
-                max-width="300"
-            >
-                <v-list density="compact">
-                    <v-list-subheader>Issue List</v-list-subheader>
-                    <template v-if="issues.length">
-                        <v-list-item
-                            v-for="(item, i) in issues"
-                            :key="i"
-                            :value="item"
+        
+        <v-col cols="12" lg="6">
+            <Card class="h-100">
+                <v-card-text>
+                    <div class="mt-4 d-flex justify-end">
+                        <v-btn
+                            @click="isShow = true"
                             color="primary"
                         >
-                            <v-list-item-title v-text="item.name"></v-list-item-title>
-                            <template v-slot:append>
-                                <div>
-                                    <v-btn 
-                                        icon="mdi-pencil-outline" 
-                                        size="small" 
-                                        @click="router.push({path: '/issue', query: {id: item.id, category_id: category_id, workspace_id: workspace_id}})">
-                                    </v-btn>
-                                    <v-btn 
-                                        icon="mdi-trash-can-outline" 
-                                        size="small" 
-                                        @click="handleDelete(item.id)">
-                                    </v-btn>
-                                </div>
-                            </template>
-                        </v-list-item>
-                    </template>
-                    <v-list-item v-else>
-                        <v-list-item-title class="text-center text-grey">
-                            No data available
-                        </v-list-item-title>
-                    </v-list-item>
-                </v-list>
-            </v-card>
+                            Add Issue
+                        </v-btn>
+                    </div>
+                    <div>
+                        <v-list>
+                            <draggable v-model="issues" tag="div" item-key="id" handle=".drag-handle" animation="300">
+                                <template #item="{ element }">
+                                <v-list-item>
+                                    <template v-slot:prepend>
+                                    <v-icon class="drag-handle" style="cursor: grab;">mdi-drag</v-icon>
+                                    </template>
+                                    <v-list-item-content>
+                                    <v-list-item-title>{{ element.name }}</v-list-item-title>
+                                    </v-list-item-content>
+                                    <template v-slot:append>
+                                        <div>
+                                            <v-btn 
+                                                icon="mdi-pencil-outline" 
+                                                size="small" 
+                                                @click="router.push({path: '/issue', query: {id: element.id, category_id: category_id, workspace_id: workspace_id}})">
+                                                </v-btn>
+                                            <v-btn 
+                                                icon="mdi-trash-can-outline" 
+                                                size="small" 
+                                                @click="handleDelete(element.id)">
+                                            </v-btn>
+                                        </div>
+                                    </template>
+                                </v-list-item>
+                                </template>
+                            </draggable>
+
+                            <!-- No Data Message inside List -->
+                            <v-list-item v-if="issues.length === 0">
+                                <v-list-item-content class="text-center">
+                                <v-list-item-title class="text-grey">No data available</v-list-item-title>
+                                </v-list-item-content>
+                            </v-list-item>
+                        </v-list>
+                    </div>
+                </v-card-text>
+            </Card>
         </v-col>
     </v-row>
     <v-dialog
@@ -145,20 +170,21 @@ onMounted(async () => {
             <v-card-title class="text-center">
                 <h5 class="text-h6 font-weight-bold">Add issue</h5>
             </v-card-title>
-            <v-card-text class="mt-5">
-                <v-row justify="center" class="align-center">
-                    <v-col cols="12">
-                        <div class="font-weight-medium mb-2 mt-5">
-                            Issue Name <i class="ph-asterisk ph-xs text-danger" />
-                        </div>
-                        <v-text-field
-                            v-model="issue_name"
-                            isRequired
-                            placeholder="Enter issue name"
-                            hide-details
-                        />
-                    </v-col>
-                </v-row>
+            <v-card-text>
+                <div class="font-weight-bold mb-2 mt-3">Name <i class="ph-asterisk ph-xs text-danger" /></div>
+                <TextField
+                    v-model="issue_name"
+                    hide-details
+                    placeholder="Enter issue name"
+                />
+                <div class="font-weight-bold mb-2 mt-3">Status <i class="ph-asterisk ph-xs text-danger" /></div>
+                <v-switch
+                    v-model="issue_status"
+                    color="primary"
+                    :label="status ? 'Enable' : 'Disable'"
+                    hide-details
+                    >
+                </v-switch>
             </v-card-text>
             <v-divider>
             </v-divider>
